refactor(scad): move FixtureParams from interfaces to fixtures module
