+++ threedframe/scad/__init__.py
@@ -1,2 +1,2 @@
 from .core import Core
-from .fixture import Fixture, SolidFixture
+from .fixture import Fixture, SolidFixture, FixtureParams
